# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CIBOL · 实验室系统 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/apikeys/mine": {
      "get": {
        "tags": [
          "apikeys"
        ],
        "summary": "My Keys",
        "operationId": "my_keys_api_apikeys_mine_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyOut"
                  },
                  "type": "array",
                  "title": "Response My Keys Api Apikeys Mine Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/apikeys": {
      "post": {
        "tags": [
          "apikeys"
        ],
        "summary": "Issue Key",
        "description": "管理员签发：把 dmxapi 后台建好的令牌下发给某成员（审批 api 请求后调用）。",
        "operationId": "issue_key_api_apikeys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyIssue"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/apikeys/{key_id}": {
      "delete": {
        "tags": [
          "apikeys"
        ],
        "summary": "Revoke Key",
        "operationId": "revoke_key_api_apikeys__key_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/booking/settings": {
      "get": {
        "tags": [
          "booking"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_api_booking_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingSettings"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "booking"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_booking_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoBookUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/booking/meetings/{meeting_id}/book": {
      "post": {
        "tags": [
          "booking"
        ],
        "summary": "Book One",
        "operationId": "book_one_api_booking_meetings__meeting_id__book_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/booking/run-auto": {
      "post": {
        "tags": [
          "booking"
        ],
        "summary": "Run Auto",
        "description": "立即触发一次自动预约扫描（便于管理员手动催一次）。",
        "operationId": "run_auto_api_booking_run_auto_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Run Auto Api Booking Run Auto Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyIssue": {
        "properties": {
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "upstream_key": {
            "type": "string",
            "title": "Upstream Key"
          },
          "budget": {
            "type": "number",
            "title": "Budget",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "user_name",
          "upstream_key"
        ],
        "title": "ApiKeyIssue"
      },
      "ApiKeyOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "masked_key": {
            "type": "string",
            "title": "Masked Key"
          },
          "budget": {
            "type": "number",
            "title": "Budget"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "used_rmb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Used Rmb"
          },
          "remain_rmb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remain Rmb"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "masked_key",
          "budget",
          "status",
          "created_at"
        ],
        "title": "ApiKeyOut"
      },
...
      "AutoBookUpdate": {
        "properties": {
          "auto_book": {
            "type": "boolean",
            "title": "Auto Book"
          }
        },
        "type": "object",
        "required": [
          "auto_book"
        ],
        "title": "AutoBookUpdate"
      },
      "BookingSettings": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "auto_book": {
            "type": "boolean",
            "title": "Auto Book"
          },
          "days_ahead": {
            "type": "integer",
            "title": "Days Ahead"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "auto_book",
          "days_ahead"
        ],
        "title": "BookingSettings"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeetingOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "y": {
            "type": "integer",
            "title": "Y"
          },
          "mo": {
            "type": "integer",
            "title": "Mo"
          },
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "mdLabel": {
            "type": "string",
            "title": "Mdlabel"
          },
          "dateLabel": {
            "type": "string",
            "title": "Datelabel"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "tone": {
            "type": "string",
            "title": "Tone"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "template": {
            "type": "string",
            "title": "Template",
            "default": "正式报告"
          },
          "scored": {
            "type": "boolean",
            "title": "Scored",
            "default": true
          },
          "host": {
            "type": "string",
            "title": "Host",
            "default": ""
          },
          "time": {
            "type": "string",
            "title": "Time",
            "default": ""
          },
          "place": {
            "type": "string",
            "title": "Place",
            "default": ""
          },
          "online": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OnlineMeetingOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "presenters": {
            "items": {
              "$ref": "#/components/schemas/PresenterOut"
            },
            "type": "array",
            "title": "Presenters"
          }
        },
        "type": "object",
        "required": [
          "id",
          "y",
          "mo",
          "day",
          "mdLabel",
          "dateLabel",
          "type",
          "tone",
          "status",
          "presenters"
        ],
        "title": "MeetingOut"
      },
...
      "OnlineMeetingOut": {
        "properties": {
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "OnlineMeetingOut"
      },
...
      "PresenterOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": ""
          },
          "minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "topic"
        ],
        "title": "PresenterOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}